Retirado espaços da escrita dos dados e inicio da leitura com seek

+++ TwitterApiRequest/Indexes/IIndexable.cs
@@ -7,6 +7,6 @@ namespace TwitterApiRequest.Indexes
     public interface IIndexable
     {
         string GetKey();
-        int GetPos();
+        long GetPos();
     }
 }
