+++ TwitterApiRequest/Indexes/IndexTweetid.cs
@@ -7,13 +7,13 @@ namespace TwitterApiRequest.Indexes
     public class IndexTweetid : IIndexable
     {
         public string Key { get; set; }
-        public int Pos { get; set; }
+        public long Pos { get; set; }
         public string GetKey()
         {
             return Key;
         }
 
-        public int GetPos()
+        public long GetPos()
         {
             return Pos;
         }
